+++ src/main/java/com/gdschongik/gdsc/domain/study/domain/StudyDetail.java
@@ -69,4 +69,8 @@ public class StudyDetail extends BaseEntity {
                 .assignment(Assignment.createEmptyAssignment())
                 .build();
     }
+
+    public void updateAssignment(Assignment newAssignment) {
+        assignment = newAssignment;
+    }
 }
